### 2019-07-18 #1582 @whedon
[PRE REVIEW]: rtdpy: A python package for residence time distributions
**Submitting author:** @MatthewFlamm (<a href="http://orcid.org/0000-0002-2040-5481">Matthew Flamm</a>)
**Repository:** <a href="https://github.com/Merck/rtdpy" target ="_blank">https://github.com/Merck/rtdpy</a>
**Version:** v0.5
