### 2019-07-18 #1582 @danielskatz
@xuanxu - if it's just 2 weeks, I think given the balance of submissions and editors, I would prefer for you to handle this one.
