### 2019-07-19 #1582 @MatthewFlamm
I've looked through the reviewer list.  Only as suggestions: both @ctdegroot and @arosen93 have chemical engineering, which would be the main applicable area for this package, listed under Domains/Topics and Python listed under Preferred Programming Language.

Look forward to the process.
