### 2019-07-23 45eda9b
Fix whitespace around references, Add caption to Figure 1
