+++ paper/paper.md
@@ -17,9 +17,9 @@ bibliography: paper.bib
 # Summary
 
 Residence time distributions (RTDs) are used to understand nonideality in chemical
-reactors and other continuous flow through steps[@levenspiel_1999]. Applications
-include chemical reaction conversion[@danckwerts_1953], pipeline mixing
-[@levenspiel_1958], and material traceability[@engisch_muzzio_2016]. RTDs also
+reactors and other continuous flow through steps [@levenspiel_1999]. Applications
+include chemical reaction conversion [@danckwerts_1953], pipeline mixing
+[@levenspiel_1958], and material traceability [@engisch_muzzio_2016]. RTDs also
 characterize flow patterns in stream flows [@haggerty_wondzell_johnson_2002] with
 applications in transport of pollutants [@kirchner_feng_neal_2000].
 
@@ -32,8 +32,8 @@ mutliple unit operations. Common RTD analyses are output prediction given an
 input signal, frequency-space signal damping, and disturbance mapping.
 
 ``rtdpy`` is a Python package that enables quick and easy computation and usage
-of RTD models.  Current model functionality is focused on continuous process
-unit operations.   RTD models include:
+of RTD models. Current model functionality is focused on continuous process
+unit operations. RTD models include:
 
 * N-continually stirred tank reactors (NCSTR), AKA tanks-in-series
 * Tube flow with Axial Dispersion
@@ -55,10 +55,10 @@ Built-in functionalities for all RTD models include:
 * Input signal convolution
 * Step response
 
-![Figure 1.](../images/ncstr.png?raw=true "N-Cstr RTDs")
+![Family of N-Cstr models.](../images/ncstr.png?raw=true "N-Cstr RTDs")
 
 ``rtdpy`` is enabled by ``numpy`` and ``scipy`` packages. Documentation on all
-major funcitonality and all RTD models is included in the repository.  Several
+major funcitonality and all RTD models is included in the repository. Several
 examples of common workflows are also provided. The base RTD class is available
 for adding additional functionality. The RTD class can also be extended for
 future or user-defined RTD models.
