### 2019-07-29 #1582 @xuanxu
:wave: @khinsen: would you be willing to review this paper? I think your knowledge of Python and computational chemistry makes you a good fit for this submission.
