### 2019-08-02 #1582 @xuanxu
:wave: @dandavies99 — would you be willing to contribute a review for JOSS on this submission?
