### 2019-08-07 #1621 @ctdegroot
A few comments as I am going through the review:

- Please add your import statements to the example given in the readme.
- I do not see any community guidelines if someone wanted to contribute to the project.
- Please add instructions to run the tests; I'm guessing pytest?
