### 2019-08-08 faebda5
Add imports to example code
