### 2019-08-08 #1621 @MatthewFlamm
Thanks @ctdegroot.  I attempted to address items 1 and 3.

For item 2, I need to get internal guidance on this and will update here once I have a path forward.
