### 2019-08-08 6c0cffd
Add Contributing
