### 2019-08-12 #1621 @ctdegroot
@MatthewFlamm Looks good. Regarding item 2, the checklist says "Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support". I think that (1) could be optional (@xuanxu please advise), but reporting issues and seeking support should be mandatory and is probably easy for you to do.
