### 2019-08-12 #1621 @MatthewFlamm
Thanks for the friendly suggestion on path forward @ctdegroot.  I'm exploring being able to satisfy the checklist to the fullest, but there is a middle way that satisfies 90% of the intent IMO.  I propose that instead of a 'Contributing to the software' section that includes point 1) from the cheklist, that this submission have a 'Extending the software' section that describes how to create a new RTD model class, for example.  But, it will not give explicit instructions on how to contribute it back to the repo, nor will it say that it cannot be contributed back. This would give detailed instructions on the 'How to develop on top of the package', but make no statement on 'How to make the PR back into the library'.

If this is not good enough, I will continue pursuing the fullest method of 'Contributing to the software' avenue.
