### 2019-08-12 #1621 @ctdegroot
@MatthewFlamm This makes sense to me and I would accept it. There is nothing stopping someone from making a PR anyways, and it would be up to you if you'd want to accept it.
