### 2019-08-13 #1621 @xuanxu
@MatthewFlamm As stated in the review criteria: `There should be clear guidelines for people wanting to contribute to the software`. Here the definition of _contribute_ is quite relaxed: proposing a new feature, finding a bug or just fixing a typo in the readme can be considered contributions. Of course that doesn't mean any of their contributions will be accepted or merged into the software, but if someone want to help there should be a clear statement on how to do it, so a user can easily find what is the _official_ preferred way (a PR, a github issue, an email?).

That could be a detailed CONTRIBUTING.md file or something as simple as a short sentence like "_If you want to contribute code, fixes, or report bugs please open a Github issue [here](https://github.com/Merck/rtdpy/issues)_",
but you have to include some pointer in the right direction for the users willing to help.
