### 2019-08-13 5d9df16
Add Contributing section to README
