### 2019-08-13 #1621 @MatthewFlamm
Thanks for the input.  I added a CONTRIBUTING.md file.
