### 2019-08-16 #1621 @dandavies99
It was a pleasure to review this software, really nice work @MatthewFlamm. Even though the application area is quite outside my own area of expertise in chemistry, I found the package easy to install, use and understand.

Like @ctdegroot, my main concern was the lack of community guidelines for contribution, but this has been dealt with now.  My only remaining, very minor suggestion would be to elaborate slightly on how tests should be carried out, and roughly what the expected output should look like. There are lots of ways to implement tests in Python and as someone who has never used pytest before, it took me a little while to work out what I was doing.

Other than that, the repo and paper both look to be in very good shape!
