### 2019-08-16 e5bf060
Add instructions to run pytest tests
