+++ README.md
@@ -31,6 +31,18 @@ See [CONTRIBUTING.md](CONTRIBUTING.md)
 ## Testing
 Tests are written using `pytest`. `numpy` and `scipy` must also be installed in the environment if using `pytest` directly. `tox` can also be used to test against Python versions 3.5, 3.6, and 3.7.
 
+It is recommended to use a virtual environment for developing/testing:
+
+```bash
+cd rtdpy_repo_path
+python3 -m venv .venv
+source .venv/bin/activate
+pip install numpy scipy pytest
+
+# run all tests
+pytest
+```
+
 Author: Matthew Flamm
 
 Email: <matthew.flamm@merck.com>
