### 2019-08-16 7c8d0a0
Add link to pytest docs and use install -e for testing instructions
