### 2019-08-16 239c87e
Add explicit instructions for getting source code for tests
