### 2019-08-16 #1621 @MatthewFlamm
Thanks @dandavies99 .  While in my experience pytest is a widely used test framework in python, I think it is a good idea to lower the barrier to entry as much as possible.  This is a good suggestion.  To this end, I've added a short section in the README.md about how to setup a development/testing environment and invoke the tests.

IMO the default output is easy to understand from pytest. I hope it suffices to link directly to the pytest documentation, which I've found to be very readable and useful.  If there was something specific you wanted clarity on in the test results output, I could add more clarification.

> to elaborate slightly on how tests should be carried out,

I had interpreted this to be about how to run the tests, not to understand the tests themselves.  Do you think the tests themselves need more documentation?
