### 2019-08-19 #1621 @dandavies99
This looks great @MatthewFlamm; exactly what was needed to lower the barrier to entry. @xuanxu I'm also done with my review now and will be very happy to see this useful package in JOSS.
