### 2019-08-19 #1621 @xuanxu
@MatthewFlamm I found a typo in the paper, Merck/rtdpy#1 should fix it.
