### 2019-08-19 cdde55e
Make acronyms and capitalization consistent
