### 2019-08-19 #1621 @MatthewFlamm
On triple reading this, I noticed a few minor things as well. I was inconsistent in capitalization in one list.  I also am missing the journal title for one reference.  I will regenerate PDF here when done.
