### 2019-08-19 #1621 @MatthewFlamm
Im still missing the journal title for the Levenspiel reference that has the title starting with Longitudinal, even though I have the journal in the bib file.  I need to read up on bib file format, I am probably missing something.
