### 2019-08-19 #1621 @xuanxu
@MatthewFlamm I can see the title of that article [in the references](https://github.com/openjournals/joss-papers/blob/joss.01621/joss.01621/10.21105.joss.01621.pdf), `Longitudinal Mixing of Fluids Flowing in Circular Pipes`, right before the other Levenspiel book.
