### 2019-08-19 #1621 @MatthewFlamm
I meant the journal name, https://github.com/Merck/rtdpy/blob/master/paper/paper.bib#L32
