### 2019-08-19 d850f6e
Fix missing journal
