+++ paper/paper.bib
@@ -29,7 +29,7 @@ title={Longitudinal Mixing of Fluids Flowing in Circular Pipes},
 volume={50},
 DOI={10.1021/ie50579a034},
 number={3},
-journal={Industrial & Engineering Chemistry},
+journal={Industrial \& Engineering Chemistry},
 author={Levenspiel, Octave},
 year={1958},
 pages={343–346}}
