### 2019-08-19 #1621 @MatthewFlamm
Fixed it, I had to escape the ampersand
