### 2019-08-19 #1621 @xuanxu
Ok, everything looks good @MatthewFlamm, here are the next steps:
- Please make a new tagged release that includes all changes made during this review
- Archive that release in Zenodo
- Check the Zenodo deposit has the correct metadata: title and author name should match the paper; you may also your ORCID.

Once you do that please report here the version number and the Zenodo DOI.
