+++ rtdpy/version.py
@@ -1,2 +1,2 @@
 """Version number."""
-__version__ = '0.5'
+__version__ = '0.5.1'
