### 2019-08-19 #1621 @MatthewFlamm
Do I need to update the version number in the paper too, if I update the version for the state of the review?
