### 2019-08-19 #1621 @xuanxu
Where do you mention the version number in the paper? I can't find it
