### 2019-08-19 #1621 @MatthewFlamm
Sorry for the confusion. For some reason I thought the version number at the top of this review issue was taken from the metadata of the paper.  You are right that the version number is nowhere in the paper.
