### 2019-08-19 #1621 @MatthewFlamm
Thanks @dandavies99 and @ctdegroot for making this open source package better.  I appreciate your comments as it is always hard to view your own creation through others eyes.
