### 2019-08-19 #1621 @xuanxu
All ready por publication :tada:. Pinging @openjournals/joss-eics for final acceptance.
