### 2019-08-20 #1621 @kyleniemeyer
Hi @MatthewFlamm, I'm nearly ready to accept this, but I noticed that in the paper Figure 1 isn't referenced/discussed in the text. Could you add a sentence or two that does this?
