### 2019-08-20 2e33961
Add generating code for Figure 1.
