### 2019-08-20 #1621 @MatthewFlamm
@kyleniemeyer I added the generating code for Figure 1 as an example usage.  The formatting now isn't my favorite with the code snippet after the figure, but I think it still all makes sense.  The code can simply be pasted in IPython, for example, and run after installing rtdpy and matplotlib.  `plt.show()` may or may not be necessary depending on the setup of the IPython console and matplotlib settings.  A basic python terminal execution would require a `plt.savefig('filename')` command, etc.

If this is all too nuanced, I'm also okay removing the code block entirely since it may not run as expected in everyone's console.  We can just leave as the explanatory text.
