### 2019-08-20 #1621 @kyleniemeyer
@MatthewFlamm I like what you added—code snippets like that are definitely encouraged when practical.
