### 2019-08-20 #1621 @kyleniemeyer
Congrats @MatthewFlamm on your article's publication in JOSS! Many thanks to @xuanxu for editing, and @ctdegroot and @dandavies99 for reviewing.
